For RestfulResource::HttpClient::ServiceUnavailable errors display the request host
Set the bugsnag notification to use the request host so in the bugsnag interface its easy to see what service is unresponsive.

Before it just showed: "HTTP 503: Service unavailable" which wasn't really helpful in say which service was down.
Now it shows: "HTTP 503: Service unavailable domain.com"

It will also use the request host to group the error notifications to reduce bugsnag spam